prod/sensitive_fields_detector_service/requirements.txt
google-cloud-bigquery
google-auth
requests
fuzzywuzzy
python-Levenshtein
networkx
matplotlib
scipy
graphviz
pandas
google-cloud-datacatalog

